***** cpmtools version 2.19 of 2014/04/08 *****
Changes since 2.18:

o  More diskdefs entries
o  New diskdefs option to set the libdsk format
o  Fixed various warnings
o  Fixed Makefile if building without curses

Originated-by: cpmtools-2.19.tar.gz
Originated-md5: e57a7add07a1a7f7b3b8196844b2aa8a
Originated-URL: http://www.moria.de/~michael/cpmtools/files/cpmtools-2.19.tar.gz

+++ cpmcp.c
@@ -259,7 +259,7 @@ int main(int argc, char *argv[])
           cpmOpen(&ino,&file,O_WRONLY);
           do
           {
-            int j;
+            unsigned int j;
 
             for (j=0; j<(sizeof(buf)/2) && (c=getc(ufp))!=EOF; ++j)
             {
@@ -267,7 +267,7 @@ int main(int argc, char *argv[])
               buf[j]=c;
             }
             if (text && c==EOF) buf[j++]='\032';
-            if (cpmWrite(&file,buf,j)!=j)
+            if (cpmWrite(&file,buf,j)!=(ssize_t)j)
             {
               fprintf(stderr,"%s: can not write %s: %s\n",cmd,dest,boo);
               ohno=1;
